Pod4 cpplint

+++ CentralComputing/tests/ConfiguratorTest.cpp
@@ -7,15 +7,15 @@ TEST(ConfiguratorTest, LoadOutputVariables) {
   EXPECT_EQ(ConfiguratorManager::config.openConfigFile("tests/test.txt"), true);
   EXPECT_EQ(ConfiguratorManager::config.openConfigFile("tests/notrealfile.txt"), false);
 
-  long long val;
+  int64_t val;
   ConfiguratorManager::config.getValue("Variable", val);
-  EXPECT_EQ(val, (long long) 12);
+  EXPECT_EQ(val, (int64_t) 12);
 
   ConfiguratorManager::config.getValue("Number", val);
-  EXPECT_EQ(val,(long long) 2342);
+  EXPECT_EQ(val, (int64_t) 2342);
 
   ConfiguratorManager::config.getValue("MaxDecel", val);
-  EXPECT_EQ(val,(long long) 231);
+  EXPECT_EQ(val, (int64_t) 231);
 
   double val2;
   ConfiguratorManager::config.getValue("MaxAccel", val2);
